Revert changes with feature and refactor together.
